feat: Implement main screen with navigation and multiple views
- Added MainScreen widget to manage navigation between Home, Publications, Explorer, and Profile views.
- Created Onboarding screens (1, 2, 3) for user onboarding process with custom buttons and text fields.
- Developed ExplorerPage with search functionality and filter chips for content categorization.
- Implemented HomeScreen to display a list of Publications using PublicationsCard widget.
- Added Profile and Publications views with basic layout.
- Created reusable widgets: CustomButton, CustomTextField, SocialButton, CodePreviewBox, and others for consistent UI.
- Integrated state management using Riverpod for navigation and filter selections.
- Enhanced UI with gradients, icons, and responsive design elements.
