+++ devinsight/lib/ui/screens/auth/registerScreen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import '../../../../config/routers/app_router.dart';
-import '../../widgets/customButton.dart';
-import '../../widgets/customTextField.dart';
+import '../../../config/routers/app_router.dart';
+import '../../widgets/auth/customButton.dart';
+import '../../widgets/auth/customTextField.dart';
 
 class RegisterScreen extends ConsumerWidget {
   const RegisterScreen({super.key});
