+++ devinsight/lib/ui/views/root/explorer.dart
@@ -1,4 +1,4 @@
-import 'package:devinsight/ui/home/widgets/searchNav.dart';
+import 'package:devinsight/ui/widgets/searchNav.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
