+++ devinsight/lib/ui/views/root/publications.dart
