+++ devinsight/lib/ui/widgets/reactionsRow.dart
@@ -1,5 +1,5 @@
 // 🔸 Fila de reacciones y compartir
-import 'package:devinsight/ui/home/widgets/ShareDialog.dart';
+import 'package:devinsight/ui/widgets/shareDialog.dart';
 import 'package:flutter/material.dart';
 
 class ReactionsRow extends StatelessWidget {
@@ -23,7 +23,8 @@ class ReactionsRow extends StatelessWidget {
       children: [
         Row(
           children: [
-            _buildReaction(Icons.check_circle, reactions[0], Colors.greenAccent),
+            _buildReaction(
+                Icons.check_circle, reactions[0], Colors.greenAccent),
             const SizedBox(width: 8),
             _buildReaction(Icons.favorite, reactions[1], Colors.yellow),
             const SizedBox(width: 8),
@@ -32,7 +33,8 @@ class ReactionsRow extends StatelessWidget {
         ),
         IconButton(
           icon: const Icon(Icons.share, color: Colors.white),
-          onPressed: () => showDialog(context: context, builder: (_) => const ShareDialog()),
+          onPressed: () =>
+              showDialog(context: context, builder: (_) => const ShareDialog()),
         )
       ],
     );
