+++ devinsight/lib/ui/widgets/searchNav.dart
